move IndexedMask to entities
