+++ scripts/utils/image.py
@@ -1,5 +1,4 @@
 from colorsys import hsv_to_rgb
-from typing import List, Tuple, Union
 
 import cv2
 import numpy as np
@@ -10,28 +9,6 @@ def gen_color_palette(n):
     rgb_array = np.array(list(map(lambda x: [int(
         v*255) for v in hsv_to_rgb(*x)], hsv_array)), dtype=np.uint8)
     return rgb_array
-
-
-class IndexedMask(np.ndarray):
-    def __new__(cls, masks: Union[np.ndarray, List, Tuple]):
-        masks = np.asarray(masks, dtype=np.uint8)
-        assert masks.ndim == 3
-        n, h, w = masks.shape
-        # cast ndarray to IndexedMask
-        self = np.zeros((h, w)).view(cls)
-        # 要検証： masksが深度降順であることを前提にしている
-        for i, mask in enumerate(masks):
-            self[mask != 0] = i + 1
-
-        self.n = n
-        self.palette = gen_color_palette(n)
-        return self
-
-    def to_rgb(self):
-        img = np.zeros((*self.shape, 3), dtype=np.uint8)
-        for i in range(self.n):
-            img[self == i+1] = self.palette[i]
-        return img
 
 
 def get_optimal_hist_th(img, n):
